build-gradle added system.properties remove testBase & testData refactor code add selenoid driver
