+++ build.gradle
@@ -46,6 +46,7 @@ tasks.withType(JavaCompile) {
 }
 
 tasks.withType(Test) {
+    systemProperties(System.getProperties())
 
     useJUnitPlatform()
