--- src/test/java/com/alexlis/helpers/DriverSettings.java
@@ -1,57 +1,0 @@
-package com.alexlis.helpers;
-
-
-import com.alexlis.config.Project;
-import com.codeborne.selenide.Configuration;
-import com.github.javafaker.App;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.remote.DesiredCapabilities;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static java.lang.String.format;
-
-public class DriverSettings {
-
-    public static void configure() {
-
-
-
-        Configuration.browser = Project.config.browser();
-        Configuration.browserVersion = Project.config.browserVersion();
-        Configuration.browserSize = Project.config.browserSize();
-//       Configuration.baseUrl = App.config.webUrl();
-
-        Configuration.remote = format("https://%s:%s@%s", login, password, browserURL);
-        Configuration.browserVersion = browserVersion;
-        Configuration.browserSize = browserSize;
-
-
-
-
-        DesiredCapabilities capabilities = new DesiredCapabilities();
-        ChromeOptions chromeOptions = new ChromeOptions();
-
-        chromeOptions.addArguments("--no-sandbox");
-        chromeOptions.addArguments("--disable-infobars");
-        chromeOptions.addArguments("--disable-popup-blocking");
-        chromeOptions.addArguments("--disable-notifications");
-        chromeOptions.addArguments("--lang=en-en");
-
-        if (Project.isWebMobile()) { // for chrome only
-            Map<String, Object> mobileDevice = new HashMap<>();
-            mobileDevice.put("deviceName", Project.config.browserMobileView());
-            chromeOptions.setExperimentalOption("mobileEmulation", mobileDevice);
-        }
-
-        if (Project.isRemoteWebDriver()) {
-            capabilities.setCapability("enableVNC", true);
-            capabilities.setCapability("enableVideo", true);
-            Configuration.remote = Project.config.remoteDriverUrl();
-        }
-
-        capabilities.setCapability(ChromeOptions.CAPABILITY, chromeOptions);
-        Configuration.browserCapabilities = capabilities;
-    }
-}
