+++ src/test/java/com/alexlis/pages/PageObjects.java
@@ -5,8 +5,6 @@ import com.alexlis.helpers.DriverUtils;
 import com.alexlis.pages.components.AuthModalForm;
 import com.codeborne.selenide.SelenideElement;
 
-import java.util.Map;
-
 import static com.codeborne.selenide.Condition.text;
 import static com.codeborne.selenide.Condition.visible;
 import static com.codeborne.selenide.Selectors.byText;
@@ -18,8 +16,6 @@ public class PageObjects {
 
     public AuthModalForm authModalForm = new AuthModalForm();
 
-//    private final String LOGIN = "Jvcr1234";
-//    private final String PASSWORD = "alexlisenkov92@mail.ru";
     private final String URL = "https://www.34play.me/";
     private final String ERROR_MESSAGE = "Логин/пароль не верен";
     private final String BANNER = "Официальный интернет магазин 34 PLAY";
@@ -208,7 +204,7 @@ public class PageObjects {
         });
     }
 
-    public void simpleCheckTitle(){
+    public void simpleCheckTitle() {
         step("Проверка соответствия текста", () -> {
             String expectedTitle = "Интернет-магазин уличной одежды 34Play";
             String actualTitle = title();
@@ -216,5 +212,4 @@ public class PageObjects {
             assertThat(actualTitle).isEqualTo(expectedTitle);
         });
     }
-
 }
